6_28

+++ utils/vqaevaluate.py
@@ -74,7 +74,7 @@ class VQAEval:
         根据公式计算出最终的Acc值
         """
         for annotation in self.answer[split]['annotations']:
-            self.acc[split][annotation['question_id']] = np.zeros( len(list(self.freq.keys())) )
+            self.acc[split][annotation['question_id']] = np.zeros( len(list(self.freq.keys())) ).astype(np.float32)
             for ans in annotation['answers']:
                 stemmed_word = self.contractions.get(ans["answer"], ans["answer"])
                 idx = self.freq.get(stemmed_word, 0)
